Update WorkloadPatch.py
Change basic call syntax of Oracle SQL*Plus from "sqlplus -s / as sysdba" to "sqlplus -S -R 2 /nolog".  The "-S" flag has the same functionality as "-s": to suppress banners and prompts.  The "-R 2" flag runs SQL*Plus in level-2 restricted mode, thus allowing SQL*Plus to start SQL script from the filesystem using the START/@ command, but it disallows the SQL*Plus commands HOST and SPOOL.  The "/nolog" flag allows SQL*Plus to successfully start even if the database instance is not running, or is running and is not OPEN.  Connecting to the database can then occur while redirecting stdin or running a SQL script.